Moved overlay.js to util
